// src/emission/enum/fuelType.enum.ts
export enum fuelType{

    LAD = 'LAD',
    LSD = 'LSD',
    LP95 = 'LP95',
    LP92 = 'LP92',
    COAL = 'Coal',
    DIESEL = 'Diesel',
    PETROL = 'PETROL',

    //boilers
    RESIDUAL_FUEL_OIL = 'RESIDUAL_FUEL_OIL',
    WOOD_WOOD_WASTE = 'WOOD_WOOD_WASTE',
    SULPHITE_LYES_BLACK_LIQUOR = 'SULPHITE_LYES_BLACK_LIQUOR',
    OTHER_PRIMARY_SOLID_BIOMASS = 'OTHER_PRIMARY_SOLID_BIOMASS',
    CHARCOAL = 'CHARCOAL',
    WOOD_CHIPS = 'WOOD_CHIPS',
    SAW_DUST = 'SAW_DUST',


    //Cooking gas
    LP_GAS ="LP_GAS",
    LANDFILLGAS = "LANDFILLGAS",
    SLUDGEGAS = "SLUDGEGAS",
    OTHERBIOGAS = "OTHERBIOGAS",
    ENERGY_INDUSTRIES = "ENERGY_INDUSTRIES",
    MANUFACTURING_INDUSTRIES_AND_CONSTRUCTION= "MANUFACTURING_INDUSTRIES_AND_CONSTRUCTION",
    COMMERCIAL_INSTITUTIONAL = "COMMERCIAL_INSTITUTIONAL",
    RESIDENTIAL_AND_AGRICULTURE_FORESTRY_FISHING_FISHING_FARM = "RESIDENTIAL_AND_AGRICULTURE_FORESTRY_FISHING_FISHING_FARM",



}