// src/emission/services/boiler/boiler.enum.ts
export enum boilerEnum {

    Furnace_Oil = "Furnace Oil",
    Solid_Biomass = "Solid Biomass",

    // ef_co2_Residual_Fuel_Oil = 77400, 
    // ef_ch4_Residual_Fuel_Oil= 3 , 
    // ef_n2o_Residual_Fuel_Oil= 0.6,
    // ncv_Residual_Fuel_Oil= 0.0404, 
    // density_Residual_Fuel_Oil = 0.97, 
    // ef_co2_Wood_Wood_Waste = 77400,  
    // ef_ch4_Wood_Wood_Waste= 3 ,
    // ef_n2o_Wood_Wood_Waste= 0.6, 
    // ncv_Wood_Wood_Waste= 0.0156, 
    // density_Wood_Wood_Waste = 1, 
    // ef_co2_Sulphite_lyes = 77400, 
    // ef_ch4_Sulphite_lyes= 3 ,
    // ef_n2o_Sulphite_lyes= 0.6,
    // ncv_Sulphite_lyes= 0.0118,
    // density_Sulphite_lyes = 1,
    // ef_co2_Other_Primary_Solid_Biomass = 77400,
    // ef_ch4_Other_Primary_Solid_Biomass= 3,
    // ef_n2o_Other_Primary_Solid_Biomass= 0.6,
    // ncv_Other_Primary_Solid_Biomass= 0.0116, 
    // density_Other_Primary_Solid_Biomass = 1, 
    // ef_co2_Charcoal = 77400, 
    // ef_ch4_Charcoal= 3,
    // ef_n2o_Charcoal= 0.6,
    // ncv_Charcoal= 0.0295, 
    // density_Charcoal = 1, 


}